refactor: update user impact and order types to use database types
