+++ src/types/index.ts
@@ -1,2 +1,2 @@
-export * from "./user";
+export * from "./database";
 export * from "./walmart";
